backend/python-nlp-service/requirements.txt
# Web Framework & Server
flask
gunicorn

# NLP Libraries
transformers
torch
spacy
nltk

# Utilities
python-dotenv